fix: evict open connections for users not present in the updated ACL

fix: evict open connections for users not present in the updated ACL


fix: evict open connections for users not present in the updated ACL


fix: evict open connections for users not present in the updated ACL


fix(server): Don't recompute shard for squashed stub tx (#2017)


feat: allow extracting the expiry time from string_set (#2020)

We introduce `iterator Find(member)` function
as well as iterator members to actually get the expiry time.

opt(server): Allow round-robin keys between shards based on prefix (#2016)

This will allow some use cases with few busy keys to distribute load
more evenly between threads.

Idea by @dranikpg.

To calculate how many entries are needed in the table I used the
following quick-n-dirty code, to reach <2.5% collision with 100 keys:

```cpp
bool Distribute(int balls = 100, int bins = 100) {
  vector<int> v(bins);
  for (int i = 0; i < balls; ++i) {
    v[rand() % v.size()]++;
  }

  for (int v : v) {
    if (v >= 2) {
      return true;
    }
  }

  return false;
}

int main(int argc, char** argv) {
  int has_2_balls = 0;
  constexpr int kRounds = 1'000'000;
  for (int i = 0; i < kRounds; ++i) {
    has_2_balls += Distribute(100, 100'000);
  }
  cout << has_2_balls << " rounds had 2+ balls in a single bin out of " << kRounds << endl;
}
```

chore: Enable squashing by default (#2011)

* chore: Enable squashing

fix: reject zset variadic commands with 0 keys (#2022)

Fixes the assertion failure as reported by #1994.


fix(server): Start periodic fiber only once init completes (#2023)

chore: remove .DS_Store file (#2025)

fix(server): Fix shardlocal eval with round-robin prefix (#2024)

fix(search): Use double numeric values (#2015)

* fix(search): Use double numeric values


---------


feat: add fieldttl command that returns the ttl of the member (#2026)

Currently implemented only for saddex.


fix(ubuntu-prod): Set suexec hash correctly (#2029)

chore(helm-chart): update to v1.11.0

fix(server): Exclude eval from pipeline squashing (#2027)

* fix(server): Exclude eval from pipeline squashing


---------


fix: evict open connections for users not present in the updated ACL

+++ src/core/search/base.h
@@ -52,7 +52,7 @@ struct WrappedStrPtr {
   std::unique_ptr<char[]> ptr;
 };
 
-using ResultScore = std::variant<std::monostate, float, int64_t, WrappedStrPtr>;
+using ResultScore = std::variant<std::monostate, float, double, WrappedStrPtr>;
 
 // Interface for accessing document values with different data structures underneath.
 struct DocumentAccessor {
